requirements.txt
# Core Flask stack
Flask==2.0.1
Flask-Bootstrap
Flask-Caching
Flask-Login==0.5.0
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.15.1
Flask-Mail==0.9.1

# Database
SQLAlchemy==1.4.23

# Forms and validation
WTForms
email-validator==1.1.3

# Security
itsdangerous
bcrypt

# Data science and export
pandas==1.3.3
numpy==1.21.2
scipy
statsmodels
markdown
weasyprint
beautifulsoup4
tenacity

# Visualization
plotly

# Other utilities
python-dotenv==0.19.0
tqdm


requests
gunicorn==20.1.0

# New additions
ydata-profiling==4.1.0

# Removed DVC-related dependencies
